+++ packages/native/src/__tests__/features/transform.test.ts
@@ -1,6 +1,7 @@
 import { describe, test, expect } from 'vitest';
 import { FastEvent } from '../../event';
-import { AssertFastMessage } from '../../types';
+import { AssertFastMessage, FastEventMessage, TransformedEvents } from '../../types';
+import { FastEventScope } from '../../scope';
 
 describe('转换事件消息', async () => {
     test('转换自定义侦听消息', () => {
@@ -143,4 +144,121 @@ describe('转换事件消息', async () => {
             });
         });
     });
+    test('在Scope中进行transform转换自定义侦听消息', () => {
+        return new Promise<void>((resolve) => {
+            type CustomEvents = TransformedEvents<{
+                'client/connect': number;
+                'client/disconnect': number;
+            }>;
+            const emitter = new FastEvent();
+
+            class MyScope extends FastEventScope<CustomEvents> {
+                constructor() {
+                    super(
+                        Object.assign(
+                            {
+                                transform: (message: FastEventMessage) => {
+                                    return message.payload;
+                                },
+                            },
+                            arguments[0],
+                        ),
+                    );
+                }
+            }
+
+            const scope = emitter.scope('div', new MyScope());
+
+            emitter.on('div/client/connect', (message) => {
+                expect(message).toBe(100);
+            });
+
+            scope.on('client/connect', (message) => {
+                expect(message).toBe(100);
+                resolve();
+            });
+            scope.emit('client/connect', 100);
+        });
+    });
+
+    test('多个Scope中进行transform转换自定义侦听消息', () => {
+        return new Promise<void>((resolve) => {
+            type ClientScopeEvents = TransformedEvents<{
+                'client/connect': number;
+                'client/disconnect': number;
+            }>;
+            type RoomScopeEvents = TransformedEvents<{
+                'room/join': { name: string };
+                'room/leabe': boolean;
+            }>;
+
+            const emitter = new FastEvent();
+
+            const clientScope = emitter.scope<ClientScopeEvents>('x');
+            clientScope.options.transform = (message) => {
+                return message.payload;
+            };
+            const roomScope = emitter.scope<RoomScopeEvents>('y');
+            roomScope.options.transform = (message) => {
+                return message.payload;
+            };
+
+            const results: any[] = [];
+            emitter.on('x/client/connect', (message) => {
+                results.push(message);
+            });
+            emitter.on('y/room/join', (message) => {
+                results.push(message);
+            });
+            clientScope.on('client/connect', (message) => {
+                expect(message).toBe(100);
+                results.push(message);
+            });
+            clientScope.emit('client/connect', 100);
+
+            roomScope.on('room/join', (message) => {
+                expect(message.name).toBe('test');
+                results.push(message);
+                expect(results).toEqual([100, 100, { name: 'test' }, { name: 'test' }]);
+                resolve();
+            });
+            roomScope.emit('room/join', { name: 'test' });
+        });
+    });
+    // test('在emitter中触发Scope中进行transform转换自定义侦听消息', () => {
+    //     return new Promise<void>((resolve) => {
+    //         type CustomEvents = TransformedEvents<{
+    //             'client/connect': number;
+    //             'client/disconnect': number;
+    //         }>;
+    //         const emitter = new FastEvent();
+
+    //         class MyScope extends FastEventScope<CustomEvents> {
+    //             constructor() {
+    //                 super(
+    //                     Object.assign(
+    //                         {
+    //                             transform: (message: FastEventMessage) => {
+    //                                 return message.payload;
+    //                             },
+    //                         },
+    //                         arguments[0],
+    //                     ),
+    //                 );
+    //             }
+    //         }
+
+    //         const scope = emitter.scope('div', new MyScope());
+
+    //         emitter.on('div/client/connect', (message) => {
+    //             expect(message.payload).toBe(100);
+    //         });
+
+    //         scope.on('client/connect', (message) => {
+    //             expect(message).toBe(100);
+    //             resolve();
+    //         });
+    //         emitter.emit('div/client/connect', 100);
+    //     });
+    // });
 });