+++ packages/native/src/scope.ts
@@ -17,15 +17,18 @@ import {
     TypedFastEventMessageOptional,
     FastEventListeners,
     RecordValues,
-    MatchEventType,
+    WildcardEvents,
     FastEventListenerFlags,
     OmitTransformedEvents,
     PickTransformedEvents,
+    FastEventMessage,
+    ClosestWildcardEvents,
 } from './types';
 import { parseEmitArgs } from './utils/parseEmitArgs';
 import { parseScopeArgs } from './utils/parseScopeArgs';
 import { renameFn } from './utils/renameFn';
 import { PickPayload, AtPayloads } from './types/index';
+import { isFunction } from './utils/isFunction';
 
 export type FastEventScopeOptions<Meta = Record<string, any>, Context = never> = {
     meta: FastEventScopeMeta & FastEventMeta & Meta;
@@ -33,6 +36,16 @@ export type FastEventScopeOptions<Meta = Record<string, any>, Context = never> =
     executor?: FastListenerExecutor;
     // 默认监听器，优先级高类方法onMessage
     onMessage?: TypedFastEventListener;
+    /**
+     * 对接收到的消息进行转换，用于将消息转换成其他格式
+     *
+     * new FastEvent({
+     *    transform:(message)=>{
+     *        message.payload
+     *    }
+     * })
+     */
+    transform?: (message: FastEventMessage) => any;
 };
 
 export interface FastEventScopeMeta {
@@ -54,7 +67,7 @@ export class FastEventScope<
         meta: undefined as unknown as FinalMeta,
         context: undefined as unknown as Fallback<Context, typeof this>,
         message: undefined as unknown as TypedFastEventMessageOptional<Events, FinalMeta>,
-        listeners: undefined as unknown as FastEventListeners<Events, FinalMeta>,
+        // listeners: undefined as unknown as FastEventListeners<Events, FinalMeta>,
         anyListener: undefined as unknown as TypedFastEventAnyListener<Events, FinalMeta, Fallback<Context, typeof this>>,
     };
     prefix: string = '';
@@ -108,13 +121,15 @@ export class FastEventScope<
         const scopeListener = renameFn(function (message: TypedFastEventMessage, args: FastEventListenerArgs) {
             const type = args.rawEventType || message.type;
             if (type.startsWith(scopePrefix)) {
-                // 消息是否经过转换
+                // 1. 判断是否经过全局Transform转换
+                // 消息是否经过转换，如果经过转换，由于结果不确定，所以不进行前缀处理，直接将消息传递给监听器
                 const isTransformed = ((args.flags || 0) & FastEventListenerFlags.Transformed) > 0;
-                const msg = isTransformed
+                let msg = isTransformed
                     ? message
                     : Object.assign({}, message, {
                           type: type.substring(scopePrefix.length),
                       });
+
                 return listener.call(scopeThis.context, msg, args);
             }
         }, listener.name) as TypedFastEventListener;
@@ -140,13 +155,13 @@ export class FastEventScope<
     // 处理使用 NotPayload 标识的事件类型
     public on<T extends keyof PickTransformedEvents<Events>>(
         type: T,
-        listener: (message: PickPayload<RecordValues<MatchEventType<Exclude<T, number | symbol>, Events>>>, args: FastEventListenerArgs<Meta>) => any | Promise<any>,
+        listener: (message: PickPayload<RecordValues<ClosestWildcardEvents<Events, Exclude<T, number | symbol>>>>, args: FastEventListenerArgs<Meta>) => any | Promise<any>,
         options?: FastEventListenOptions<Events, Meta>,
     ): FastEventSubscriber;
 
     public on<T extends Exclude<string, Types>>(
         type: T,
-        listener: TypedFastEventAnyListener<MatchEventType<T, Events>, Meta, Fallback<Context, typeof this>>,
+        listener: TypedFastEventAnyListener<ClosestWildcardEvents<Events, T>, Meta, Fallback<Context, typeof this>>,
         options?: FastEventListenOptions,
     ): FastEventSubscriber;
     public on(type: '**', listener: TypedFastEventAnyListener<Events, FinalMeta, Fallback<Context, typeof this>>, options?: FastEventListenOptions): FastEventSubscriber;
@@ -171,13 +186,13 @@ export class FastEventScope<
     // 处理使用 NotPayload 标识的事件类型
     public once<T extends keyof PickTransformedEvents<Events>>(
         type: T,
-        listener: (message: PickPayload<RecordValues<MatchEventType<Exclude<T, number | symbol>, Events>>>, args: FastEventListenerArgs<Meta>) => any | Promise<any>,
+        listener: (message: PickPayload<RecordValues<WildcardEvents<Events, Exclude<T, number | symbol>>>>, args: FastEventListenerArgs<Meta>) => any | Promise<any>,
         options?: FastEventListenOptions<Events, Meta>,
     ): FastEventSubscriber;
 
     public once<T extends Exclude<string, Types>>(
         type: T,
-        listener: TypedFastEventAnyListener<MatchEventType<T, Events>, Meta, Fallback<Context, typeof this>>,
+        listener: TypedFastEventAnyListener<WildcardEvents<Events, T>, Meta, Fallback<Context, typeof this>>,
         options?: FastEventListenOptions,
     ): FastEventSubscriber;
     public once(): FastEventSubscriber {
@@ -214,10 +229,10 @@ export class FastEventScope<
     public emit(type: Types, directive: symbol): void;
     public emit(type: string, directive: symbol): void;
     public emit<R = any, T extends Types = Types>(type: T, payload?: PickPayload<Events[T]>, retain?: boolean): R[];
-    public emit<R = any, T extends string = string>(type: T, payload?: PickPayload<T extends Types ? Events[T] : RecordValues<MatchEventType<T, Events>>>, retain?: boolean): R[];
+    public emit<R = any, T extends string = string>(type: T, payload?: PickPayload<T extends Types ? Events[T] : RecordValues<WildcardEvents<Events, T>>>, retain?: boolean): R[];
     public emit<R = any, T extends string = string>(
         type: T,
-        payload?: PickPayload<T extends Types ? Events[T] : RecordValues<MatchEventType<T, Events>>>,
+        payload?: PickPayload<T extends Types ? Events[T] : RecordValues<WildcardEvents<Events, T>>>,
         options?: FastEventListenerArgs<FinalMeta>,
     ): R[];
     public emit<R = any>(message: FastEventEmitMessage<AtPayloads<Events>, FinalMeta>, options?: FastEventListenerArgs<FinalMeta>): R[];
@@ -235,10 +250,22 @@ export class FastEventScope<
         if (arguments.length === 2 && typeof arguments[0] === 'string' && arguments[1] === FastEventDirectives.clearRetain) {
             return this.emitter.emit(this._getScopeType(arguments[0])!);
         }
-        const [message, options] = parseEmitArgs(arguments, this.emitter.options.meta, this.options.meta, this.options.executor);
+        let [message, args] = parseEmitArgs(arguments, this.emitter.options.meta, this.options.meta, this.options.executor);
+
+        this._transformMessage(message, args);
         message.type = this._getScopeType(message.type)!;
-        return this.emitter.emit(message as TypedFastEventMessage<Events, FinalMeta>, options);
+
+        return this.emitter.emit(message as TypedFastEventMessage<Events, FinalMeta>, args);
     }
+
+    private _transformMessage(message: FastEventMessage, args: FastEventListenerArgs<FinalMeta>) {
+        if (isFunction(this._options.transform)) {
+            args.rawEventType = this._getScopeType(message.type)!;
+            args.flags = (args.flags || 0) | FastEventListenerFlags.Transformed;
+            message.payload = this._options.transform.call(this, message);
+        }
+    }
+
     public async emitAsync<R = any, T extends Types = Types>(type: T, payload?: PickPayload<Events[T]>, retain?: boolean): Promise<[R | Error][]>;
     public async emitAsync<R = any, T extends string = string>(type: T, payload?: PickPayload<T extends Types ? Events[T] : any>, retain?: boolean): Promise<[R | Error][]>;
     public async emitAsync<R = any, T extends Types = Types>(type: T, payload?: PickPayload<Events[T]>, options?: FastEventListenerArgs<FinalMeta>): Promise<[R | Error][]>;
@@ -320,7 +347,7 @@ export class FastEventScope<
         P extends string = string,
         M extends Record<string, any> = Record<string, any>,
         C = Context,
-        ScopeInstance extends FastEventScope<any, any, any> = FastEventScope<ScopeEvents<Events, P> & E, FinalMeta & M, C>,
+        ScopeInstance extends FastEventScope<Record<string, any>, any, any> = FastEventScope<Record<string, any>, any, any>,
     >(
         prefix: P,
         scopeObj: ScopeInstance,